chore: Split React components into different files
chore: Update next config
