+++ components/deployed-tab/index.tsx
@@ -1,108 +1,10 @@
-import { FaRegClipboard } from "react-icons/fa";
-import { MdDone } from "react-icons/md"
 import styles from './style.module.css';
 
-import { smartdeploy, FetchDatas } from "@/pages";
-import { Ok, Err } from 'smartdeploy-client'
-import { useState, useEffect, Dispatch, SetStateAction } from "react";
+import { FetchDatas } from "@/pages";
+import { DeployedContract, listAllDeployedContracts } from './smartdeploy-functions';
+import { useState, useEffect } from "react";
+import ClipboardIconComponent from './clip-board-component';
 import { useThemeContext } from '../ThemeContext'
-
-interface DeployedContract {
-    index: number;
-    name: string;
-    address: string;
-}
-
-type ClipboardIconComponentProps = {
-    address: string;
-}
-
-
-async function listAllDeployedContracts() {
-
-    try {
-
-        ///@ts-ignore
-        const { result } = await smartdeploy.listDeployedContracts({ start: undefined, limit: undefined });
-        const response = result;
-
-        if (response instanceof Ok) {
-            
-            let deployedContracts: DeployedContract[] = [];
-
-            const contractArray =  response.unwrap();
-
-            ///@ts-ignore
-            contractArray.forEach(([name, address], i) => {
-
-                const parsedDeployedContract: DeployedContract = {
-                    index: i,
-                    name: name,
-                    address: address.toString(),
-                }
-
-                deployedContracts.push(parsedDeployedContract);
-                
-            });
-            
-            //console.log(deployedContracts);
-            return deployedContracts;
-
-        } else if (response instanceof Err) {
-            response.unwrap();
-        } else {
-            throw new Error("listDeployedContracts returns undefined. Impossible to fetch the deployed contracts.");
-        }
-    } catch (error) {
-        console.error(error);
-        window.alert(error);
-    }
-    
-}
-
-async function copyAddr(setCopied: Dispatch<SetStateAction<boolean>> , addr: string) {
-    await navigator.clipboard
-                             .writeText(addr)
-                             .then(() => {
-                                setCopied(true);
-                             })
-                             .catch((err) => {
-                                console.error("Failed to copy address: ", err);
-                                window.alert(err);
-                             });
-}
-
-function ClipboardIconComponent(props: ClipboardIconComponentProps) {
-
-    const [copied, setCopied] = useState<boolean>(false);
-  
-    useEffect(() => {
-        if(copied === true) {
-          const timer = setTimeout(() => {
-            setCopied(false)
-          }, 1500);
-          return () => clearTimeout(timer);
-        }
-    }, [copied]);
-  
-    return (
-        <>
-            {!copied ? (
-                <td className={styles.clipboardIconCell}>
-                    <FaRegClipboard 
-                        className={styles.clipboardIcon}
-                        onClick={ () => copyAddr(setCopied, props.address)}
-                    />
-                </td>
-            ) : (
-                <td className={styles.clipboardIconCell}>
-                    <p className={styles.copiedMessage}><MdDone style={{ marginRight: '0.2rem' }}/>Copied!</p>
-                </td>
-            )}
-        </>
-    );
-}
-
 
 export default function DeployedTab(props: FetchDatas) {
