+++ next.config.js
@@ -1,9 +1,6 @@
 module.exports = {
   output: "export",
   basePath: process.env.NODE_ENV === "production" ? "": undefined,
-  experimental: {
-    appDir: true,
-  },
   images: {
     unoptimized: true,
   },
